I have added a Posts.js container, and have edited the blog file to only render Posts component.  Later on I will render the route to Full Post.
